added min-ins and max-ins random ins

+++ DataTypes/Process.h
@@ -10,8 +10,8 @@
 class Process {
     public:
         std::string name;
-        int current_instruction;
-        int total_instructions;
+        long long current_instruction;
+        long long total_instructions;
         std::string timestamp;
         bool completed;
         int core;
@@ -21,7 +21,7 @@ class Process {
 
         Process() {}
 
-        Process(std::string name, int total_instructions, std::string timestamp) {
+        Process(std::string name, long long total_instructions, std::string timestamp) {
             this->name = name;
             this->id = last_id++;;
             this->current_instruction = 0;
